Pass all tests without fence_i, div* and rem*

+++ run-all-tests.sh
@@ -4,13 +4,14 @@ set -eu
 
 # Clone riscv-tests and install to riscv-tests/target before running this script.
 
-rm -rf tests-tmp
-mkdir tests-tmp
-cp riscv-tests/target/share/riscv-tests/isa/rv32ui-p-* tests-tmp
-cp riscv-tests/target/share/riscv-tests/isa/rv32um-p-* tests-tmp
-rm tests-tmp/*.dump
+if [ ! -e tests-tmp ]; then
+  mkdir tests-tmp
+  cp riscv-tests/target/share/riscv-tests/isa/rv32ui-p-* tests-tmp
+  cp riscv-tests/target/share/riscv-tests/isa/rv32um-p-* tests-tmp
+  rm tests-tmp/*.dump
+fi
 
-for test in $(ls ./tests-tmp);
+for test in $(ls ./tests-tmp | grep -v ".*\.bin");
 do
   ./run-one-test.sh $test nowin
 done
